comments are much nicer now closes #195 + voting/commenting does not refresh full video page anymore closes #179
