+++ lib/bloc/postdetails/postdetails_repository.dart
@@ -5,14 +5,14 @@ import 'package:http/http.dart' as http;
 import 'dart:convert';
 
 abstract class PostRepository {
-  Future<Post> getPost(
-      String apiNode, String author, String link, String applicationUser);
+  Future<Post> getPost(String apiNode, String author, String link,
+      String applicationUser, bool topLevel);
 }
 
 class PostRepositoryImpl implements PostRepository {
   @override
   Future<Post> getPost(String apiNode, String author, String link,
-      String applicationUser) async {
+      String applicationUser, bool topLevel) async {
     Post _post = Post(
         sId: "",
         author: "",
@@ -22,8 +22,8 @@ class PostRepositoryImpl implements PostRepository {
         tags: [],
         isFlaggedByUser: false);
 
-    String _tag = "";
-    while (_tag == "") {
+    if (!topLevel) {
+      // load exactly this post / comment
       var response = await http.get(Uri.parse(apiNode +
           AppConfig.postDataUrl
               .replaceAll("##AUTHOR", author)
@@ -32,22 +32,40 @@ class PostRepositoryImpl implements PostRepository {
         var _data = json.decode(response.body);
 
         Post _post = ApiResultModel.fromJson(_data, applicationUser).post;
-        _tag = _post.jsonString!.tag!;
-        if (_tag == "") {
-          author = _post.pa!;
-          link = _post.pp!;
-        } else {
-          return _post;
-        }
+
+        return _post;
       } else {
-        print("post other exception");
         throw Exception();
       }
-    }
-    if (_post.sId != "") {
-      return _post;
     } else {
-      throw Exception();
+      // iterate through potential comments to top level post
+      String _tag = "";
+      while (_tag == "") {
+        var response = await http.get(Uri.parse(apiNode +
+            AppConfig.postDataUrl
+                .replaceAll("##AUTHOR", author)
+                .replaceAll("##LINK", link)));
+        if (response.statusCode == 200) {
+          var _data = json.decode(response.body);
+
+          Post _post = ApiResultModel.fromJson(_data, applicationUser).post;
+          _tag = _post.jsonString!.tag!;
+          if (_tag == "") {
+            author = _post.pa!;
+            link = _post.pp!;
+          } else {
+            return _post;
+          }
+        } else {
+          print("post other exception");
+          throw Exception();
+        }
+      }
+      if (_post.sId != "") {
+        return _post;
+      } else {
+        throw Exception();
+      }
     }
   }
 }
