+++ lib/style/ThemeData.dart
@@ -110,80 +110,80 @@ final ThemeData dtubeDarkTheme = ThemeData(
   textTheme: TextTheme(
     headline1: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 30,
+      fontSize: 25,
       fontWeight: FontWeight.w900,
       fontStyle: FontStyle.normal,
     ),
     headline2: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 30,
+      fontSize: 23,
       fontWeight: FontWeight.w700,
       fontStyle: FontStyle.normal,
     ),
     headline3: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 25,
+      fontSize: 21,
       fontWeight: FontWeight.w900,
       fontStyle: FontStyle.normal,
     ),
     headline4: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 25,
+      fontSize: 19,
       fontWeight: FontWeight.w700,
       fontStyle: FontStyle.normal,
     ),
     headline5: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 20,
+      fontSize: 17,
       fontWeight: FontWeight.w700,
       fontStyle: FontStyle.normal,
     ),
     headline6: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 17,
+      fontSize: 15,
       fontWeight: FontWeight.w700,
       fontStyle: FontStyle.normal,
     ),
     subtitle1: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 17,
+      fontSize: 15,
       fontWeight: FontWeight.w200,
       fontStyle: FontStyle.normal,
     ),
     bodyText1: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 17,
+      fontSize: 15,
       fontWeight: FontWeight.w400,
       fontStyle: FontStyle.normal,
     ),
     bodyText2: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 15,
+      fontSize: 13,
       fontWeight: FontWeight.w400,
       fontStyle: FontStyle.normal,
     ),
     caption: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 15,
+      fontSize: 13,
       fontWeight: FontWeight.w400,
       fontStyle: FontStyle.normal,
     ),
     button: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 17,
+      fontSize: 15,
       fontWeight: FontWeight.w400,
       fontStyle: FontStyle.normal,
     ),
     subtitle2: GoogleFonts.workSans(
       color: globalTextColor,
-      fontSize: 14,
+      fontSize: 13,
       fontWeight: FontWeight.w400,
       fontStyle: FontStyle.normal,
     ),
     overline: GoogleFonts.workSans(
       decoration: TextDecoration.underline,
       color: globalTextColor,
-      fontSize: 17,
+      fontSize: 13,
       fontWeight: FontWeight.w400,
       fontStyle: FontStyle.normal,
     ),
